Updated the custom channel setting, joining, and notification for others
I added the ability to set a custom channel name as follows:
Set the channel to 'custom' and type the custom chat channel name into the 'bet gold' box. The first run of a new game, with the custom channel selected will pull from the custom chat channel name you set in the box. Until reloading / logging out, you'll be able to stay locked in that channel, even if you change the channel mode, or change the box input.